# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI Todo App",
    "description": "A simple todo app built with FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/user/me": {
      "get": {
        "tags": ["user"],
        "summary": "User Profile",
        "operationId": "user_profile_user_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/token"
          }
        }
      }
    }
  }
}